pgml-apps/pgml-chat/pyproject.toml
[tool.poetry]
name = "pgml-chat"
version = "0.1.1"
description = "PostgresML bot builder for all your documentation"
authors = ["PostgresML <team@postgresml.org>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pgml_chat"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
openai = "^0.27.8"
rich = "^13.4.2"
pgml = "^0.9.0"
python-dotenv = "^1.0.0"
click = "^8.1.6"
black = "^23.7.0"
slack-bolt = "^1.18.0"
discord-py = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pgml-chat = "pgml_chat.main:main"
